decode: truncate the 32-bit reads now that V's `int` is 64-bit
vlang/v#28293 made `int` 64-bit on 64-bit targets, which silently removed
the truncation read_int32/read_sfixed32 were relying on. sfixed32 -1 came
back as 4294967295 — caught by the protoc oracle, which byte-matched on
encode and then disagreed on decode.

Both reads now go through an explicit i32 cast, so 32-bit semantics come
from the code rather than from the width of `int`. read_sint32 already
truncated to u32 before zigzag, and the 64-bit reads are unaffected.

The fuzz generator relied on the same implicit truncation to stay inside
int32 for an int32 field.

+++ decoder.v
@@ -127,9 +127,11 @@ pub fn (mut d Decoder) read_string() !string {
 	return d.read_view()!.bytestr()
 }
 
-// int32/int64 arrive sign-extended to 64 bits; truncation restores them.
+// int32/int64 arrive sign-extended to 64 bits; truncation restores them. The
+// i32 cast is what truncates: `int` is 64-bit on 64-bit targets since
+// vlang/v#28293, so casting straight to it would keep the upper bits.
 pub fn (mut d Decoder) read_int32() !int {
-	return int(d.read_varint()!)
+	return int(i32(d.read_varint()!))
 }
 
 pub fn (mut d Decoder) read_int64() !i64 {
@@ -156,7 +158,8 @@ pub fn (mut d Decoder) read_sint64() !i64 {
 }
 
 pub fn (mut d Decoder) read_sfixed32() !int {
-	return int(d.read_fixed32()!)
+	// via i32 so the sign bit survives the widening to a 64-bit `int`
+	return int(i32(d.read_fixed32()!))
 }
 
 pub fn (mut d Decoder) read_sfixed64() !i64 {
